Add support for STLink-V3 and STLink V2.1
Add support for the STLinkV3. This device uses the same protocol as the
STLinkV2 just uses a different endpoint.

The STLink V2.1 can be found on newer Nucleo boards, it uses the same
endpoint as the STLinkV3.

Many thanks to Antonio Borneo (who contributed the STLinkV3 support to
the OpenOCD) for the help and hints!