Fix text output in stm8flash to display immediately
When running stm8flash from within another IDE (such as Code::Blocks), any text that is output does not get displayed in the IDE until stm8flash exits (due to a lack of a flush). This commit fixes this by wrapping the text output calls in a macro that flushes the output stream each time. With this fix, the text output is displayed as stm8flash is running, and you are able to see it's progress in real time.