main: fix unlock mechanism for stm8l chipsets
Fix the mechanism to remove the readout protection (ROP) of an STM8L
chipset.

ST's document PM0054 describes the unlock procedure in section 4.1:
"To unprotect the device, the ROP must be written twice: the first write,
with any value, launches a global erase which includes the option byte;
the second write, 0xAA, unprotects the device, and then the 0xAA value
remains in the ROP byte."

The definition of ROP_STM8L has a comment that matches the description in
PM0054. It also says that we have to write 0xAA into the ROP option byte
twice.

My tests showed that this double 0xAA write actually sets the ROP option
byte to 0xAA and resets all other option bytes to their default values.
There's no need to clear the other option bytes manually.
part->option_bytes_size is now unused for ROP_STM8L...