espstlink: Flash blocks less than the device's block size
If writing a block to be flashed, an entire block must be sent or it
will not be written.

Using the code from stlinkv2.c, espstlink reads the existing data
in the area to be flashed (up to the end of the last affected flash
block) and add the existing data to the end of the new data if
necessary.